requirements.txt
pyserial
crcmod-1.7